requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
beautifulsoup4==4.12.3
black
bs4==0.0.2
colorama==0.4.0
constantly==15.1.0
cryptography==42.0.6
dataclasses-json==0.6.5
diskcache==5.6.3
dnspython==2.6.1
einops==0.7.0
environs==9.5.0
exceptiongroup==1.2.1
fastapi==0.110.0
Flask==2.0.0
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.4.0
frozenlist==1.4.1
fsspec==2024.3.1
g4f==0.3.0.8
gguf==0.6.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
isort==5.13.2
itemadapter==0.8.0
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-text-splitters==0.0.1
langsmith==0.1.54
llama_cpp_python==0.2.75
logparser==0.8.2
MarkupSafe==2.0.0
marshmallow==3.21.2
minio==7.2.7
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
orjson==3.10.3
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
pexpect==4.7.0
pillow==10.3.0
platformdirs==4.2.1
protobuf==5.26.1
ptyprocess==0.6.0
pyarrow==16.0.0
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyDispatcher==2.0.5
pymilvus==2.3.6
pymongo==4.6.3
pyngrok==7.1.6
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==21.0.1
pyvi==0.1.1
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2024.4.28
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
scrapyd==1.4.3
scrapyd-client==1.2.3
sentence-transformers==2.7.0
sentencepiece==0.1.99
sklearn-crfsuite==0.3.6
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.5.0
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.4
transformers==4.38.2
triton==2.1.0
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==1.5.1
uberegg==0.1.1
ujson==5.9.0
uvicorn==0.28.1
w3lib==2.0.0
Werkzeug==2.0.0
yarl==1.9.4
